Move LanguageModal to src/components/navigation
